+++ display/tab.go
@@ -3,6 +3,7 @@ package display
 import (
 	"fmt"
 	"net/url"
+	"path"
 	"strconv"
 	"strings"
 
@@ -20,19 +21,31 @@ const (
 	tabModeLoading
 )
 
+// tabHistoryPageCache is fields from the Page struct, cached here to solve #122
+// See structs/structs.go for an explanation of the fields.
+type tabHistoryPageCache struct {
+	row        int
+	column     int
+	selected   string
+	selectedID string
+	mode       structs.PageMode
+}
+
 type tabHistory struct {
-	urls []string
-	pos  int // Position: where in the list of URLs we are
+	urls      []string
+	pos       int // Position: where in the list of URLs we are
+	pageCache []*tabHistoryPageCache
 }
 
 // tab hold the information needed for each browser tab.
 type tab struct {
-	page     *structs.Page
-	view     *cview.TextView
-	history  *tabHistory
-	mode     tabMode
-	barLabel string // The bottomBar label for the tab
-	barText  string // The bottomBar text for the tab
+	page             *structs.Page
+	view             *cview.TextView
+	history          *tabHistory
+	mode             tabMode
+	barLabel         string // The bottomBar label for the tab
+	barText          string // The bottomBar text for the tab
+	preferURLHandler bool   // For #143, use URL handler over proxy
 }
 
 // makeNewTab initializes an tab struct with no content.
@@ -86,6 +99,7 @@ func makeNewTab() *tab {
 			linkN, _ := strconv.Atoi(currentSelection[0])
 			tabs[tab].page.Selected = tabs[tab].page.Links[linkN]
 			tabs[tab].page.SelectedID = currentSelection[0]
+			tabs[tab].preferURLHandler = false // Reset in case
 			go followLink(tabs[tab], tabs[tab].page.URL, tabs[tab].page.Links[linkN])
 			return
 		}
@@ -204,11 +218,24 @@ func makeNewTab() *tab {
 				return nil
 			}
 			return nil
+		case config.CmdURLHandlerOpen:
+			currentSelection := t.view.GetHighlights()
+			t.preferURLHandler = true
+			// Copied code from when enter key is pressed
+			if len(currentSelection) > 0 {
+				bottomBar.SetLabel("")
+				linkN, _ := strconv.Atoi(currentSelection[0])
+				t.page.Selected = t.page.Links[linkN]
+				t.page.SelectedID = currentSelection[0]
+				go followLink(&t, t.page.URL, t.page.Links[linkN])
+			}
+			return nil
 		}
 		// Number key: 1-9, 0, LINK1-LINK10
 		if cmd >= config.CmdLink1 && cmd <= config.CmdLink0 {
 			if int(cmd) <= len(t.page.Links) {
 				// It's a valid link number
+				t.preferURLHandler = false // Reset in case
 				go followLink(&t, t.page.URL, t.page.Links[cmd-1])
 				return nil
 			}
@@ -313,6 +340,21 @@ func makeNewTab() *tab {
 	return &t
 }
 
+// historyCachePage caches certain info about the current page in the tab's history,
+// see #122 for details.
+func (t *tab) historyCachePage() {
+	if t.page == nil || t.page.URL == "" || t.history.pageCache == nil || len(t.history.pageCache) == 0 {
+		return
+	}
+	t.history.pageCache[t.history.pos] = &tabHistoryPageCache{
+		row:        t.page.Row,
+		column:     t.page.Column,
+		selected:   t.page.Selected,
+		selectedID: t.page.SelectedID,
+		mode:       t.page.Mode,
+	}
+}
+
 // addToHistory adds the given URL to history.
 // It assumes the URL is currently being loaded and displayed on the page.
 func (t *tab) addToHistory(u string) {
@@ -320,9 +362,15 @@ func (t *tab) addToHistory(u string) {
 		// We're somewhere in the middle of the history instead, with URLs ahead and behind.
 		// The URLs ahead need to be removed so this new URL is the most recent item in the history
 		t.history.urls = t.history.urls[:t.history.pos+1]
+		// Same for page cache
+		t.history.pageCache = t.history.pageCache[:t.history.pos+1]
 	}
 	t.history.urls = append(t.history.urls, u)
 	t.history.pos++
+
+	// Cache page info for #122
+	t.history.pageCache = append(t.history.pageCache, &tabHistoryPageCache{}) // Add new spot
+	t.historyCachePage()                                                      // Fill it with data
 }
 
 // pageUp scrolls up 75% of the height of the terminal, like Bombadillo.
@@ -380,7 +428,7 @@ func (t *tab) applyHorizontalScroll() {
 		// Scrolled to the right far enough that no left margin is needed
 		browser.AddTab(
 			strconv.Itoa(i),
-			makeTabLabel(strconv.Itoa(i+1)),
+			t.label(),
 			makeContentLayout(t.view, 0),
 		)
 		t.view.ScrollTo(t.page.Row, t.page.Column-leftMargin())
@@ -388,7 +436,7 @@ func (t *tab) applyHorizontalScroll() {
 		// Left margin is still needed, but is not necessarily at the right size by default
 		browser.AddTab(
 			strconv.Itoa(i),
-			makeTabLabel(strconv.Itoa(i+1)),
+			t.label(),
 			makeContentLayout(t.view, leftMargin()-t.page.Column),
 		)
 	}
@@ -498,4 +546,36 @@ func (t *tab) highlightedURL() string {
 		return selectedURL
 	}
 	return ""
+}
+
+// label returns the label to use for the tab name
+func (t *tab) label() string {
+	tn := tabNumber(t)
+	if tn < 0 {
+		// Invalid tab, shouldn't happen
+		return ""
+	}
+
+	// Increment so there's no tab 0 in the label
+	tn++
+
+	if t.page.URL == "" || t.page.URL == "about:newtab" {
+		// Just use tab number
+		// Spaces around to keep original Amfora look
+		return fmt.Sprintf(" %d ", tn)
+	}
+	if strings.HasPrefix(t.page.URL, "about:") {
+		// Don't look for domain, put the whole URL except query strings
+		return strings.SplitN(t.page.URL, "?", 2)[0]
+	}
+	if strings.HasPrefix(t.page.URL, "file://") {
+		// File URL, use file or folder as tab name
+		return path.Base(t.page.URL[7:])
+	}
+	// Otherwise, it's a Gemini URL
+	pu, err := url.Parse(t.page.URL)
+	if err != nil {
+		return fmt.Sprintf(" %d ", tn)
+	}
+	return pu.Host
 }